refactor: Move isUpDownMetric file
