+++ src/GmdVizPanel/statushistory/isUpDownMetric.ts
